Modularize medic area schemas
