+++ app/api/medic_area/turns.py
@@ -1,0 +1,329 @@
+"""Turn management routes."""
+from typing import List, Optional, TypeVar
+from uuid import UUID
+
+from fastapi import APIRouter, Depends, HTTPException, Request, status
+from fastapi.responses import ORJSONResponse
+from sqlmodel import select
+
+from app.db.main import SessionDep
+from app.models import (
+    Appointments,
+    Departments,
+    Doctors,
+    Specialties,
+    Turns,
+    User,
+)
+from app.schemas.medica_area import (
+    AppointmentResponse,
+    DepartmentResponse,
+    DoctorResponse,
+    PayTurnResponse,
+    ServiceResponse,
+    SpecialtyResponse,
+    TurnsCreate,
+    TurnsDelete,
+    TurnsResponse,
+    TurnsState,
+)
+from app.core.interfaces.medic_area import TurnAndAppointmentRepository
+from app.core.services.stripe_payment import StripeServices
+
+from .common import auth_dependency, console, default_response_class
+
+
+router = APIRouter(
+    prefix="/turns",
+    tags=["turns"],
+    default_response_class=default_response_class,
+    dependencies=[auth_dependency()],
+)
+
+
+@router.get("/", response_model=List[TurnsResponse])
+async def get_turns(request: Request, session: SessionDep):
+    if not request.state.user.is_superuser:
+        raise HTTPException(status_code=401, detail="You are not authorized")
+
+    statement = select(Turns)
+    result = session.exec(statement).all()
+
+    turns_serialized: List[TurnsResponse] = []
+    for turn in result:
+        turns_serialized.append(
+            TurnsResponse(
+                id=turn.id,
+                reason=turn.reason,
+                state=turn.state,
+                date=turn.date,
+                date_limit=turn.date_limit,
+                date_created=turn.date_created,
+                user_id=turn.user_id,
+                doctor_id=turn.doctor_id,
+                appointment_id=str(turn.appointment.id),
+                time=turn.time,
+                service=[
+                    ServiceResponse(
+                        id=serv.id,
+                        name=serv.name,
+                        description=serv.description,
+                        price=serv.price,
+                        specialty_id=serv.specialty_id,
+                    )
+                    for serv in turn.services
+                ],
+            ).model_dump()
+        )
+
+    return ORJSONResponse(turns_serialized)
+
+
+Serializer = TypeVar("Serializer")
+
+
+def serialize_model(
+    model: object, serializer: Serializer, session, refresh: bool = False
+) -> Serializer:
+    """Helper to copy fields from a model into a serializer instance."""
+    fields = serializer.__fields__.keys()
+    serializer_instance = serializer()
+    for field in fields:
+        setattr(serializer_instance, field, getattr(model, field))
+
+    if refresh:
+        session.refresh(model)
+
+    return serializer_instance
+
+
+@router.get("/{user_id}", response_model=Optional[List[TurnsResponse]])
+async def get_turns_by_user_id(
+    request: Request, session: SessionDep, user_id: UUID
+):
+    user = session.get(User, user_id)
+
+    def serialize_departament(department: Departments):
+        session.merge(department)
+        session.refresh(department)
+        return DepartmentResponse(
+            id=department.id,
+            name=department.name,
+            description=department.description,
+            location_id=department.location_id,
+        )
+
+    def serialize_speciality(speciality: Specialties):
+        session.merge(speciality)
+        session.refresh(speciality)
+        return SpecialtyResponse(
+            id=speciality.id,
+            name=speciality.name,
+            description=speciality.description,
+            department_id=speciality.department_id,
+            department=serialize_departament(speciality.departament),
+        )
+
+    try:
+        turns_serialized = [
+            TurnsResponse(
+                id=turn.id,
+                reason=turn.reason,
+                state=turn.state,
+                date=turn.date,
+                date_limit=turn.date_limit,
+                date_created=turn.date_created,
+                user_id=turn.user_id,
+                time=turn.time,
+                doctor=DoctorResponse(
+                    id=turn.doctor.id,
+                    dni=turn.doctor.dni,
+                    username=turn.doctor.name,
+                    speciality_id=turn.doctor.speciality_id,
+                    date_joined=turn.doctor.date_joined,
+                    is_active=turn.doctor.is_active,
+                    email=turn.doctor.email,
+                    first_name=turn.doctor.first_name,
+                    last_name=turn.doctor.last_name,
+                    telephone=turn.doctor.telephone,
+                ).model_dump(),
+                service=[
+                    ServiceResponse(
+                        id=service.id,
+                        name=service.name,
+                        description=service.description,
+                        price=service.price,
+                        specialty_id=service.specialty_id,
+                    ).model_dump()
+                    for service in turn.services
+                ],
+            ).model_dump()
+            for turn in user.turns
+        ]
+
+        return ORJSONResponse(turns_serialized, status_code=200)
+    except Exception as exc:  # pragma: no cover - mirrors behaviour
+        console.print_exception(show_locals=True)
+        raise HTTPException(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, detail=str(exc)
+        ) from exc
+
+
+@router.get("/{turn_id}", response_model=TurnsResponse)
+async def get_turn_by_id(request: Request, session: SessionDep, turn_id: UUID):
+    user = request.state.user
+
+    try:
+        secure_turn = session.exec(select(Turns).where(Turns.id == turn_id)).first()
+
+        if secure_turn is None:
+            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Turn not found")
+
+        if secure_turn.user_id != user.id or secure_turn.doctor_id != user.id:
+            raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Unauthorized")
+
+        return ORJSONResponse(
+            TurnsResponse(
+                id=secure_turn.id,
+                reason=secure_turn.reason,
+                state=secure_turn.state,
+                date=secure_turn.date,
+                date_limit=secure_turn.date_limit,
+                date_created=secure_turn.date_created,
+                user_id=secure_turn.user_id,
+            ).model_dump()
+        )
+    except Exception as exc:  # pragma: no cover - behaviour preserved
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=str(exc)) from exc
+
+
+@router.post("/add", response_model=PayTurnResponse)
+async def create_turn(request: Request, session: SessionDep, turn: TurnsCreate):
+    user: User | None = request.state.user
+
+    if "doc" in request.state.scopes:
+        if turn.user_id is None:
+            raise HTTPException(status_code=400, detail="user_id is required")
+
+    elif user:
+        turn.user_id = user.id
+
+    else:
+        raise HTTPException(status_code=500, detail="Internal Error")
+
+    try:
+        new_turn, new_appointment = await TurnAndAppointmentRepository.create_turn_and_appointment(
+            session=session,
+            turn=turn,
+        )
+
+        if not new_turn:
+            raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail=new_appointment)
+
+        response = await StripeServices.proces_payment(
+            price=new_turn.price_total(),
+            details=new_turn.get_details(),
+            h_i=turn.health_insurance,
+            session=session,
+        )
+
+        return ORJSONResponse(
+            PayTurnResponse(
+                turn=TurnsResponse(
+                    id=new_turn.id,
+                    reason=new_turn.reason,
+                    state=new_turn.state,
+                    date=new_turn.date,
+                    date_limit=new_turn.date_limit,
+                    date_created=new_turn.date_created,
+                    user_id=new_turn.user_id,
+                    time=new_turn.time,
+                ).model_dump(),
+                payment_url=response,
+            ).model_dump(),
+            status_code=status.HTTP_201_CREATED,
+        )
+
+    except HTTPException as exc:
+        console.print_exception(show_locals=True)
+        raise HTTPException(status_code=exc.status_code, detail=exc.detail)
+
+    except Exception as exc:  # pragma: no cover - keep behaviour
+        raise HTTPException(status_code=400, detail=str(exc)) from exc
+
+
+@router.delete("/delete/{turn_id}", response_model=TurnsDelete)
+async def delete_turn(request: Request, session: SessionDep, turn_id: UUID):
+    session_user = request.state.user
+    try:
+        turn = session.get(Turns, turn_id)
+        if session_user.id != turn.user_id or session_user.id != turn.doctor_id:
+            raise HTTPException(status_code=403, detail="You are not authorized")
+        deleted = TurnAndAppointmentRepository.delete_turn_and_appointment(session, turn)
+        if not deleted:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail=f"the turn: {turn_id} cannot be deleted",
+            )
+        return ORJSONResponse(
+            TurnsDelete(
+                id=turn.id,
+                message=f"Turn {turn.id} has been deleted",
+            ),
+            status_code=status.HTTP_200_OK,
+        )
+    except Exception as exc:  # pragma: no cover - preserve behaviour
+        raise HTTPException(status_code=400, detail=str(exc)) from exc
+
+
+@router.patch("/update/state")
+async def update_state(request: Request, turn_id: UUID, new_state: str, session: SessionDep):
+    turn = session.get(Turns, turn_id)
+
+    if not turn:
+        raise HTTPException(404, detail="Turn Not Found")
+
+    if "superuser" not in request.state.scopes:
+        if turn.state.value in ["finished", "rejected", "cancelled"]:
+            raise HTTPException(
+                status_code=status.HTTP_409_CONFLICT,
+                detail="Turn has a not mutable state",
+            )
+
+    try:
+        match new_state:
+            case "waiting":
+                turn.state = TurnsState.waiting
+
+            case "finished":
+                turn.state = TurnsState.finished
+
+            case "cancelled":
+                turn.state = TurnsState.cancelled
+
+            case "rejected":
+                turn.state = TurnsState.rejected
+
+            case "accepted":
+                turn.state = TurnsState.accepted
+
+            case _:
+                raise HTTPException(
+                    status.HTTP_501_NOT_IMPLEMENTED,
+                    detail=f"{new_state} isn´t in the valid states",
+                )
+
+        session.add(turn)
+        session.commit()
+
+        return ORJSONResponse({"msg": "success"}, status_code=200)
+
+    except HTTPException:
+        raise
+
+    except Exception as exc:  # pragma: no cover - retain behaviour
+        console.print_exception(show_locals=True)
+        raise HTTPException(500, detail=str(exc)) from exc
+
+
+__all__ = ["router"]
